+++ src/data/Type.js
@@ -69,18 +69,4 @@ module.exports = class extends Type {
 
     return transformers;
   }
-
-  getResolvers() {
-    const resolvers = [];
-    const scalarType = this.field.getScalarRef();
-
-    if (scalarType) {
-      Object.entries(scalarType.getDirectiveArgs('field', {})).forEach(([key, value]) => {
-        if (!Array.isArray(value)) value = [value];
-        if (key === 'resolve') resolvers.push(...value.map(t => Transformer.getInstances()[t]));
-      });
-    }
-
-    return resolvers;
-  }
 };