+++ src/graphql/ast/Field.js
@@ -44,27 +44,6 @@ module.exports = class Field extends Node {
     return this.getDirectiveArg('field', 'default');
   }
 
-  resolveBoundValue(query, initialValue) {
-    // If no bound value, return default value
-    const defaultValue = uvl(initialValue, this.getDefaultValue());
-    if (!this.hasBoundValue()) return defaultValue;
-
-    // Grab @value definition, if passive then check for initialValue
-    const { scope, path, passive = false } = this.getDirectiveArgs('value');
-    if (passive && initialValue !== undefined) return initialValue;
-
-    // Resolve @value
-    switch (scope) {
-      case 'context': {
-        const { resolver } = query.toObject();
-        const context = resolver.getContext();
-        const value = get(context, path);
-        return uvl((typeof value === 'function') ? value() : value, defaultValue);
-      }
-      default: return this[path];
-    }
-  }
-
   // Model Methods
   getSchema() {
     return this.model.getSchema();