+++ src/graphql/ast/Node.js
@@ -2,7 +2,6 @@ const { get } = require('lodash');
 const { Kind } = require('graphql');
 const { nvl, uvl } = require('../../service/app.service');
 const { mergeAST } = require('../../service/graphql.service');
-// const Memoizer = require('../../data/Memoizer');
 
 const operations = ['Query', 'Mutation', 'Subscription'];
 const modelKinds = [Kind.OBJECT_TYPE_DEFINITION, Kind.OBJECT_TYPE_EXTENSION, Kind.INTERFACE_TYPE_DEFINITION, Kind.INTERFACE_TYPE_EXTENSION];
@@ -18,7 +17,6 @@ module.exports = class Node {
     this.toString = () => this.getName();
     this.nodeType = nodeType;
     this.name = get(this.ast, 'name.value');
-    // return new Memoizer(this, Object.getOwnPropertyNames(Node.prototype).filter(m => ['getContext'].indexOf(m) === -1));
   }
 
   // Basic AST Methods
@@ -164,13 +162,6 @@ module.exports = class Node {
    */
   isVirtual() {
     return Boolean(this.getDirectiveArg('link', 'by'));
-  }
-
-  /**
-   * Does the model/field have a bound @value directive
-   */
-  hasBoundValue() {
-    return Boolean(this.getDirective('value'));
   }
 
   /**